refactor: extract getCartByCustomerId
